backend/requirements.txt
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.1.1
Werkzeug>=2.0
click>=7.1.2
itsdangerous>=2.0
Jinja2>=3.0
boto3==1.38.38
Six
sqlalchemy-libsql
python-dotenv
